THz pulse in time LB in space
Note that it is only possible to define one optical penetration depth for all the layers. 
That is, not like in sourcetype = "Gaussian", where different penetration depths with respect to the layer can be given, here only one, for all the layers can be considered, that is a smooth decay all over space.